+++ Cargo.toml
@@ -7,9 +7,8 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
-clap = "3.1"
+clap = { version = "3.1", features = ["derive"] }
 fuel-types = "0.3"
-fuels-signers = "0.10"
 fuel-crypto = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
